refactor: Introducing TokenType KeywordWithDollarString
A keyword followed by a dollar string is a valid identifier in QBasic. For example `DIM$ ="hello"` creates a new variable named `DIM$`, even though `DIM` is a keyword.

This doesn't happen with the rest of the type qualifiers (!, #, %, &).

To skip having to handle this everywhere in the parsing logic, moving this logic into the recognizers.