refactor: Simplified implicits in linter
No longer passing along the collected implicits all the way. Added an implicits variable in Names (should be in Context in future revision). There is actually only one place in the code where an implicit is being added, and that place now invokes the add function of the implicits directly.