refactor: Introducing Stateful trait
Inspired by reading about the State monad. Although I don't fully understand it, this is an attempt to applying some inspirations.

The Stateful trait is effectively implementing the shared mutable state, however, the implementation looks more like the Environment (or Reader) monad, because the signature is (&mut State) -> Value.

Started applying it in a few places in the Converter, it's already helping with not having to pass the state around from function to function.