fix: Linter was reporting a valid variable as illegal
Variable names that contain a dot are illegal if the part before the dot is the name of a user defined variable declared anywhere in the program.

For example:

```
DIM A AS SomeUserDefinedType
DIM A.B
```

In this case `A.B` is illegal because it starts with `A`, which is of a user defined type.

The linter would incorrectly report as illegal cases where the variable `A` was of any extended type (e.g. `AS INTEGER`).