refactor: Remove Rc from Context
Context only needs a reference to a PreLinterResult. Storing that reference to the struct was done with a Rc.

Switching to a real reference like &'a causes problems as every state type, like DimListState, ExprState, etc, will need an additional reference parameter. Same for the Convertible trait, whose default C type is Context. So using an &'a reference is not practical.

The easiest approach (a compromise) is to let Context own the PreLinterResult and consume it back at the end of the conversion operation.