refactor: Simplify converter for statements
The converter of statements would convert a `Statement` to an
`Option<Statement>`. The only case where it would return `None`
was the `CONST` statement, as the `CONST` statement is not needed
after the linter is done.

Keeping the `CONST` in there and ignoring it is not harmful either
though, so it's better to do that, at the benefit of having
a simplified mapping function.