refactor: Use `VariableInfo` from linter `Context` in generator
The `VariableInfo` in `Expression` is not really used in the parser,
except for its `expression_type` field.

As a first step of removing `VariableInfo` from the `Expression`
in the parser, we're exposing the Linter's `Context` (which
has this information) to the Instruction Generator.

There, the Instruction Generator can get the necessary information
from the Linter's Context, instead of the `Expression`.