feat(pc): Introducing context for parsers
Adding a new generic type to the `Parser` trait,
`C`, which by default will be an empty tuple `()`.

The intention is to use this (in a later commit)
in order to support more complex parsing scenarios
in an expressive way, where the parsing
depends on what has been parsed so far.
Currently that is done with the `then_with`
parser but it is too verbose and requires
propagating the context manually.